+++ src/js/file/document.js
@@ -5,15 +5,6 @@ function bufferToBase64(buf) {
   return btoa(binstr);
 }
 
-function base64ToBuffer(base64) {
-  var binstr = atob(base64);
-  var buf = new Uint8Array(binstr.length);
-  Array.prototype.forEach.call(binstr, function (ch, i) {
-    buf[i] = ch.charCodeAt(0);
-  });
-  return buf;
-}
-
 function isString (value) {
   return typeof value === 'string' || value instanceof String;
 }